requirements.txt
pandas==2.2.1
openpyxl==3.1.2
google-api-python-client==2.111.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1
PyDrive2==1.16.4
